+++ index.html
@@ -6,7 +6,7 @@
   <link rel="stylesheet" href="css/normalize.css">
   <link rel="stylesheet" href="css/style.css">
   <link rel="shortcut icon" href="favicon.ico" type="image/x-icon">
-  <script src="functions.js"></script>
+
 
   <title>Кекстаграм</title>
 </head>
@@ -27,65 +27,6 @@
     <!-- Контейнер для изображений от других пользователей -->
     <section class="pictures  container">
       <h2 class="pictures__title  visually-hidden">Фотографии других пользователей</h2>
-
-      <a href="#" class="picture">
-        <img class="picture__img" src="./photos/1.jpg" width="182" height="182" alt="Случайная фотография">
-        <p class="picture__info">
-          <span class="picture__comments"></span>
-          <span class="picture__likes"></span>
-        </p>
-      </a>
-
-      <a href="#" class="picture">
-        <img class="picture__img" src="./photos/2.jpg" width="182" height="182" alt="Случайная фотография">
-        <p class="picture__info">
-          <span class="picture__comments"></span>
-          <span class="picture__likes"></span>
-        </p>
-      </a>
-
-      <a href="#" class="picture">
-        <img class="picture__img" src="./photos/3.jpg" width="182" height="182" alt="Случайная фотография">
-        <p class="picture__info">
-          <span class="picture__comments"></span>
-          <span class="picture__likes"></span>
-        </p>
-      </a>
-      <a href="#" class="picture">
-        <img class="picture__img" src="./photos/4.jpg" width="182" height="182" alt="Случайная фотография">
-        <p class="picture__info">
-          <span class="picture__comments"></span>
-          <span class="picture__likes"></span>
-        </p>
-      </a>
-      <a href="#" class="picture">
-        <img class="picture__img" src="./photos/5.jpg" width="182" height="182" alt="Случайная фотография">
-        <p class="picture__info">
-          <span class="picture__comments"></span>
-          <span class="picture__likes"></span>
-        </p>
-      </a>
-      <a href="#" class="picture">
-        <img class="picture__img" src="./photos/6.jpg" width="182" height="182" alt="Случайная фотография">
-        <p class="picture__info">
-          <span class="picture__comments"></span>
-          <span class="picture__likes"></span>
-        </p>
-      </a>
-      <a href="#" class="picture">
-        <img class="picture__img" src="./photos/7.jpg" width="182" height="182" alt="Случайная фотография">
-        <p class="picture__info">
-          <span class="picture__comments"></span>
-          <span class="picture__likes"></span>
-        </p>
-      </a>
-      <a href="#" class="picture">
-        <img class="picture__img" src="./photos/8.jpg" width="182" height="182" alt="Случайная фотография">
-        <p class="picture__info">
-          <span class="picture__comments"></span>
-          <span class="picture__likes"></span>
-        </p>
-      </a>
 
       <!-- Поле для загрузки нового изображения на сайт -->
       <section class="img-upload">
@@ -294,6 +235,6 @@
         <h2 class="data-error__title">Не удалось загрузить данные</h2>
       </section>
     </template>
-    <script src="./js/main.js"></script>
+    <script src="js/main.js" type="module"></script>
 </body>
 </html>
